refactor: move Python module templates to external files that can be linted (#6105)
* Python modules: move templates to external files that can be linted


* Fix linter


* Split unrelated bug fix into another PR


---------
