+++ internal/mage/util/util.go
@@ -32,7 +32,6 @@ func Repository(c *dagger.Client) *dagger.Directory {
 			"**/.pytest_cache",
 			"**/.ruff_cache",
 			"sdk/python/dist",
-			"sdk/python/experimental",
 
 			// go
 			// go.work is ignored so that you can use ../foo during local dev and let
