+++ sdk/python/pyproject.toml
@@ -142,7 +142,7 @@ target-version = ["py310", "py311"]
 
 [tool.ruff]
 src = ["src", "tests"]
-exclude = ["experimental/*/pyproject.toml"]
+ignore-init-module-imports = true
 target-version = "py310"
 select = ["ALL"]
 ignore = [
@@ -188,12 +188,9 @@ unfixable = [
 [tool.ruff.isort]
 known-first-party = ["dagger"]
 
-[tool.ruff.flake8-bugbear]
-extend-immutable-calls = ["typer.Option"]
-
 [tool.ruff.per-file-ignores]
 "./docs/*" = ["D1"]
-"./experimental/*" = ["D1", "T201"]
+"./runtime/*" = ["D1"]
 "./src/dagger/client/gen.py" = [
     # Not much control over field names and docs coming from the API.
     # Note: We could detect built-in shadowing like the reserved
@@ -228,7 +225,7 @@ extend-immutable-calls = ["typer.Option"]
     "UP007",
 ]
 # Allow some patterns to redefine imports in __init__.
-"__init__.py" = ["I001", "F401", "F403", "PLC0414"]
+"./src/**/__init__.py" = ["I001", "F403", "PLC0414"]
 # Typer uses boolean parameters for the CLI. Let it.
 "./src/**/cli.py" = ["FBT"]
