+++ sdk/python/runtime/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	_ "embed"
 	"os"
 	"path"
 	"path/filepath"
@@ -11,46 +12,15 @@ type PythonSdk struct{}
 const (
 	ModSourceDirPath      = "/src"
 	RuntimeExecutablePath = "/runtime"
-	sdkSrc                = "/sdk"
 	venv                  = "/opt/venv"
+	sdkSrc                = "/sdk"
 	genDir                = "sdk"
 	genPath               = "src/dagger/client/gen.py"
+	schemaPath            = "/schema.json"
 )
 
-var pyprojectTmpl = `[project]
-name = "main"
-version = "0.0.0"
-`
-
-var srcMainTmpl = `import dagger
-from dagger.mod import function
-
-
-@function
-def container_echo(string_arg: str) -> dagger.Container:
-    # Example usage: "dagger call container-echo --string-arg hello"
-    return dagger.container().from_("alpine:latest").with_exec(["echo", string_arg])
-
-
-@function
-async def grep_dir(directory_arg: dagger.Directory, pattern: str) -> str:
-    # Example usage: "dagger call grep-dir --directory-arg . --patern grep_dir"
-    return await (
-        dagger.container()
-        .from_("alpine:latest")
-        .with_mounted_directory("/mnt", directory_arg)
-        .with_workdir("/mnt")
-        .with_exec(["grep", "-R", pattern, "."])
-        .stdout()
-    )
-`
-
-var runtimeTmpl = `#!/usr/bin/env python
-import sys
-from dagger.mod.cli import app
-if __name__ == '__main__':
-    sys.exit(app())
-`
+//go:embed scripts/runtime.py
+var runtimeTmpl string
 
 func (m *PythonSdk) ModuleRuntime(modSource *Directory, subPath string, introspectionJson string) *Container {
 	return m.CodegenBase(modSource, subPath, introspectionJson).
@@ -83,27 +53,28 @@ func (m *PythonSdk) Codegen(modSource *Directory, subPath string, introspectionJ
 
 func (m *PythonSdk) CodegenBase(modSource *Directory, subPath string, introspectionJson string) *Container {
 	return m.Base("").
+		WithDirectory(sdkSrc, dag.Host().Directory(root(), HostDirectoryOpts{
+			Exclude: []string{"runtime"},
+		})).
+		WithMountedDirectory("/opt", dag.Host().Directory(root(), HostDirectoryOpts{
+			Include: []string{"runtime/template"},
+		})).
+		WithExec([]string{"python", "-m", "pip", "install", "-e", sdkSrc}).
 		WithMountedDirectory(ModSourceDirPath, modSource).
 		WithWorkdir(path.Join(ModSourceDirPath, subPath)).
-		// TODO: Move all of this to a python script.
-		WithNewFile("/templates/pyproject.toml", ContainerWithNewFileOpts{
-			Contents: pyprojectTmpl,
-		}).
-		WithNewFile("/templates/src/main.py", ContainerWithNewFileOpts{
-			Contents: srcMainTmpl,
-		}).
-		WithNewFile("/schema.json", ContainerWithNewFileOpts{
+		WithNewFile(schemaPath, ContainerWithNewFileOpts{
 			Contents: introspectionJson,
 		}).
+		// TODO: Move all of this to a python script, add more intelligence.
 		WithExec([]string{
 			"python", "-m", "dagger", "codegen",
 			"--output", path.Join(sdkSrc, genPath),
-			"--introspection", "/schema.json",
+			"--introspection", schemaPath,
 		}, ContainerWithExecOpts{
 			ExperimentalPrivilegedNesting: true,
 		}).
-		WithExec([]string{"sh", "-c", "[ -f pyproject.toml ] || cp /templates/pyproject.toml ."}).
-		WithExec([]string{"sh", "-c", "find . -name '*.py' | grep -q . || { mkdir -p src; cp /templates/src/main.py src/main.py; }"})
+		WithExec([]string{"sh", "-c", "[ -f pyproject.toml ] || cp /opt/runtime/template/pyproject.toml ."}).
+		WithExec([]string{"sh", "-c", "find . -name '*.py' | grep -q . || { mkdir -p src; cp /opt/runtime/template/src/main.py src/main.py; }"})
 }
 
 func (m *PythonSdk) Base(version string) *Container {
@@ -117,11 +88,7 @@ func (m *PythonSdk) Base(version string) *Container {
 		WithEnvVariable("VIRTUAL_ENV", venv).
 		WithEnvVariable("PATH", "$VIRTUAL_ENV/bin:$PATH", ContainerWithEnvVariableOpts{
 			Expand: true,
-		}).
-		WithDirectory(sdkSrc, dag.Host().Directory(root(), HostDirectoryOpts{
-			Exclude: []string{"runtime"},
-		})).
-		WithExec([]string{"python", "-m", "pip", "install", "-e", sdkSrc})
+		})
 }
 
 // TODO: fix .. restriction
