+++ sdk/python/runtime/scripts/runtime.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+
+import sys
+
+from dagger.mod.cli import app
+
+if __name__ == "__main__":
+    sys.exit(app())
