+++ sdk/python/runtime/template/pyproject.toml
@@ -1,0 +1,3 @@
+[project]
+name = "main"
+version = "0.0.0"
